west: custom command for downloading in Codespaces
Use the custom `west download` command when working in Codespaces to
automatically add the app name and timestamp to the binary and move it
to the /workspace/Downloads folder where it is easily found.

Once in this folder, binaries may be downloaded by right-clicking on the
filename and selecting "download".
